Clean up CopyButton and move to experimental
